+++ HISTORY.md
@@ -86,6 +86,7 @@
 ### 2019-05-31
 * Threads: optimized FOREACH performance for large linked lists
 * German translation update (FINDFIRST, WORD, TEMPORARY?)
+* Threads: also allow numbers as sprite names in SET
 
 ### 2019-05-29
 * Threads, Objects: added "Find First" primitive to lists category